Update patch set 1
Patch Set 1: -Code-Review

Well, I'm afraid we'll get in the same situation with relying on a new version of PyPI ("This is a pre-production deployment of Warehouse" says its web site). I think we're going to be much safer using tarballs.o.o.

Not blocking this patch though.

Patch-set: 1
Label: Code-Review=0