Update patch set 3
Patch Set 3: Code-Review-2

uhm this went wrong, I'd prefer we restore PS1 to match what is already in newton-rdo to keep it in sync.
Then as a followup change we could modify Source0 to use tarballs.o.o but only if this package is not in Fedora, otherwise I'm not sure if https://fedoraproject.org/wiki/Upstream_release_monitoring will work.

Patch-set: 3